utils/jwt_utils.py
from fastapi import HTTPException
from jwt.exceptions import DecodeError
from datetime import datetime, timedelta, UTC
import jwt

from config import SECRET_KEY, ALGORITHM

def create_access_token(data: dict, expires_delta: timedelta = None):
    to_encode = data.copy()
    if expires_delta:
        expire = datetime.now(UTC) + expires_delta
    else:
        expire = datetime.now(UTC) + timedelta(minutes=30)
    to_encode.update({"exp": expire})
    try:
        encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
    except DecodeError as e:
        raise HTTPException(status_code=500, detail="Failed to create access token") from e
    return encoded_jwt

class JWTBearer:
    def verify_jwt(self, jwt_token: str):
        try:
            payload = jwt.decode(jwt_token, SECRET_KEY, algorithms=[ALGORITHM])
            return payload
        except jwt.ExpiredSignatureError:
            raise HTTPException(status_code=401, detail="Token expired")
        except jwt.InvalidTokenError:
            raise HTTPException(status_code=401, detail="Invalid token")